@@ -1,52 +1,80 @@
 {
-  "name": "she-means-business",
+  "name": "smb-financial-manager",
   "main": "src/App.js",
   "version": "1.0.0",
   "scripts": {
     "start": "expo start",
     "reset-project": "node ./scripts/reset-project.js",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
+    "android": "expo run:android",
+    "ios": "expo run:ios",
     "web": "expo start --web",
     "test": "jest --watchAll",
     "lint": "expo lint"
   },
+  "jest": {
+    "preset": "react-native",
+    "transform": {
+      "^.+\\.jsx?$": "babel-jest",
+      "^.+\\.tsx?$": "babel-jest"
+    },
+    "transformIgnorePatterns": [
+      "node_modules/(?!(react-native|my-project|react-native-button|@react-native|@react-navigation|@react-native-community|@react-native-picker|@react-native-async-storage)/)"
+    ]
+  },
   "dependencies": {
+    "@expo/image-utils": "^0.6.3",
+    "@expo/prebuild-config": "^8.0.17",
     "@expo/vector-icons": "^14.0.4",
+    "@react-native-async-storage/async-storage": "1.23.1",
     "@react-native-community/datetimepicker": "^8.2.0",
-    "@react-native-picker/picker": "^2.7.5",
+    "@react-native-firebase/app": "^21.4.0",
+    "@react-native-firebase/auth": "^21.4.0",
+    "@react-native-firebase/database": "^21.4.0",
+    "@react-native-picker/picker": "2.9.0",
     "@react-navigation/native": "^6.1.18",
     "@react-navigation/stack": "^6.4.1",
-    "expo": "~51.0.0",
-    "expo-constants": "~16.0.2",
-    "expo-font": "~12.0.10",
-    "expo-image-picker": "~15.0.7",
-    "expo-linking": "~6.3.1",
-    "expo-router": "~3.5.23",
-    "expo-splash-screen": "^0.27.6",
-    "expo-status-bar": "~1.12.1",
-    "expo-system-ui": "~3.0.7",
-    "expo-web-browser": "~13.0.3",
-    "lottie-react-native": "^6.7.0",
+    "date-fns": "^4.1.0",
+    "expo": "^52.0.7",
+    "expo-build-properties": "^0.13.1",
+    "expo-constants": "~17.0.3",
+    "expo-file-system": "^18.0.3",
+    "expo-font": "~13.0.1",
+    "expo-image-picker": "~16.0.2",
+    "expo-linking": "~7.0.2",
+    "expo-modules-autolinking": "^2.0.1",
+    "expo-router": "^4.0.6",
+    "expo-sharing": "^13.0.0",
+    "expo-splash-screen": "~0.29.11",
+    "expo-status-bar": "~2.0.0",
+    "expo-system-ui": "~4.0.3",
+    "expo-web-browser": "~14.0.1",
+    "firebase": "^11.0.1",
+    "lottie-react-native": "^7.1.0",
     "lucide-react-native": "^0.454.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-native": "^0.74.5",
+    "pdf-lib": "^1.17.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-native": "0.76.2",
     "react-native-chart-kit": "^6.12.0",
-    "react-native-gesture-handler": "~2.16.1",
-    "react-native-reanimated": "~3.10.1",
-    "react-native-safe-area-context": "4.10.5",
-    "react-native-screens": "3.31.1",
-    "react-native-svg": "15.2.0",
+    "react-native-gesture-handler": "~2.20.2",
+    "react-native-reanimated": "~3.16.1",
+    "react-native-safe-area-context": "4.12.0",
+    "react-native-screens": "~4.1.0",
+    "react-native-svg": "15.8.0",
     "react-native-walkthrough-tooltip": "^1.6.0",
     "react-native-web": "~0.19.10",
     "recharts": "^2.13.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.0",
-    "@types/react": "~18.2.45",
+    "@babel/core": "^7.26.0",
+    "@babel/preset-env": "^7.26.0",
+    "@babel/preset-react": "^7.25.9",
+    "@babel/preset-typescript": "^7.26.0",
+    "@testing-library/react-native": "^12.8.1",
+    "@types/react": "~18.3.12",
+    "babel-jest": "^29.7.0",
     "jest": "^29.2.1",
-    "jest-expo": "~51.0.4",
+    "jest-expo": "~52.0.2",
     "typescript": "~5.3.3"
   },
   "private": true
